+++ internal/diagnose/config_validator.go
@@ -19,8 +19,6 @@ import (
 
 const (
 	validationEventType          = "NrIntegrationError"
-	DefaultPostRetryDelaySec     = 5
-	DefaultPostMaxRetries        = 20
 	DefaultMaxValidationAttempts = 20
 )
 
@@ -45,8 +43,8 @@ func NewConfigValidator(client *newrelic.NewRelic) *ConfigValidator {
 	return &ConfigValidator{
 		client:               client,
 		PollingNRQLValidator: v,
-		PostRetryDelaySec:    DefaultPostRetryDelaySec,
-		PostMaxRetries:       DefaultPostMaxRetries,
+		PostRetryDelaySec:    config.DefaultPostRetryDelaySec,
+		PostMaxRetries:       config.DefaultPostMaxRetries,
 	}
 }
 
@@ -59,7 +57,7 @@ func (c *ConfigValidator) Validate(ctx context.Context) error {
 
 	i, err := host.InfoWithContext(ctx)
 	if err != nil {
-		log.Error(err)
+		log.Debug(err)
 		return ErrDiscovery
 	}
 
@@ -72,7 +70,7 @@ func (c *ConfigValidator) Validate(ctx context.Context) error {
 
 	postEvent := func() error {
 		if err = c.client.Events.CreateEventWithContext(ctx, accountID, evt); err != nil {
-			log.Error(err)
+			log.Debug(err)
 			return ErrPostEvent
 		}
 
@@ -94,7 +92,7 @@ func (c *ConfigValidator) Validate(ctx context.Context) error {
 	`, evt.EventType, evt.Hostname, evt.GUID)
 
 	if _, err = c.PollingNRQLValidator.Validate(ctx, query); err != nil {
-		log.Error(err)
+		log.Debug(err)
 		err = ErrValidation
 	}