+++ CHANGELOG.md
@@ -1,6 +1,31 @@
-<a name="v0.31.0"></a>
-## [v0.31.0] - 2021-07-20
+<a name="v0.31.4"></a>
+## [v0.31.4] - 2021-07-23
+<a name="v0.31.3"></a>
+## [v0.31.3] - 2021-07-23
 ### Bug Fixes
+- failure events
+
+<a name="v0.31.2"></a>
+## [v0.31.2] - 2021-07-22
+### Bug Fixes
+- emit canceled status correctly
+- install tutone before generating
+- increase retries when fetching license key
+- **install:** lower log level to debug when retry failures occur during validation
+- **logging:** pass logrus instance to NR client
+
+<a name="v0.31.1"></a>
+## [v0.31.1] - 2021-07-21
+### Bug Fixes
+- **execution:** use better interface contract to match nerdgraph API
+
+### Features
+- **install:** update log level to debug when failing to update status
+
+<a name="v0.31.0"></a>
+## [v0.31.0] - 2021-07-21
+### Bug Fixes
+- **install:** fix loading embedded files on windows
 - **install:** suppress spinner output during verbose logging
 - **install:** remove spinner suppression during debug logging
 - **install:** obfuscate license key
@@ -904,7 +929,11 @@
 - **profile:** Enable reading of profiles and use Region/APIKey from default profile
 - **profile:** Add listing of profiles to command
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.0...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.4...HEAD
+[v0.31.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.3...v0.31.4
+[v0.31.3]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.2...v0.31.3
+[v0.31.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.1...v0.31.2
+[v0.31.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.31.0...v0.31.1
 [v0.31.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.30.5...v0.31.0
 [v0.30.5]: https://github.com/newrelic/newrelic-client-go/compare/v0.30.4...v0.30.5
 [v0.30.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.30.3...v0.30.4